Send access token in Fabric calls
This is the TypeScript side of work to support the token-based
authentication and authorization workflow in communications over
Fabric.

By default, Bifrost will look for a token under key `csp-auth-token` in
sessionStorage and will inject it in the header of a STOMP message
so that the Bifrost running in the backend will be able to validate it.

This MR also adds a couple utility functions.

**fabric.getAccessTokenSessionStorageKey()**
Get the key used to retrieve the access token from sessionStorage

**fabric.setAccessTokenSessionStorageKey(key: string)**
Set the key used to retrieve the access token from sessionStorage

**fabric.getAccessToken()**
Get the token from sessionStorage