Fix lints and command chain causing lints to slip
The MR replaces all double quotes with single quotes for
the files that were recently modified for galactic store support.
It also fixes an issue with the "build-docs" npm script that was
chaining `gulp npm:prepare` with `npm run build-docs` with `;`
which would not consider the exit status of the gulp command.
This effectively allowed any lint errors reported as shown in
CI pipeline logs to slip by and allow the rest of the gulp pipeline
to continue.

In this particular case, since linting was part of the TypeScript
compilation stage, the TS compilation task failed as a whole, and
as a result, there were no JS artifacts at the end of the chain.