Expose global HTTP headers for remote REST service
Today, we define and use the headers store in Bifrost TS's
RestService (a.k.a local REST service). This makes it less ideal for
SewingMachine-generated operation classes to reach the store can get
global HTTP headers. This MR adds a getter method to FabricApiImpl
to expose HTTP global headers, so that the operation class methods can
easily access the global HTTP headers and pass them through the
web socket to the Bifrost Java's RestService.