Enforce use of typescript version 3.2
The latest release 0.14.24 was built with typescript@3.6.3 which,
by default, makes use of `globalThis` in type definition files.
`globalThis` is only recognized by typescript library version of 3.4
and higher, which is a problem for projects depending on Angular 7
because Angular 7 strictly allows a typescript versions range of
>=3.1.1 and <3.3. To keep support for Angular 7 we will need to
stick with typescript library of version 3.2.4 which is the latest
release for 3.2.x.