Fix lost/improper galactic message routing in concurrent
This MR fixes an issue where multiple galactic messages to the
same channel are routed to wrong `listen()` handlers or even lost.
This issue only happens when these events happen at almost the same
time window where the response for a request could be returned to
the handler of another request operation.