1.1 Patch from gitlab
Patch 1/2 from gitlab for v1.1

Patching over from internal gitlab.

* Bump up version from package.json and add relnote for 1.1.0
* Add comment to connString based on Dave's suggestion
* Clean up code to get connection string

Instead of constructing connection string multiple times in
various places throughout the code base having each invocation call the
singleton method that works the same across the board ensures easier
maintenance of the code. The reason the existing
`BusUtil.getFabricConnectionString()` was deprecated in this commit is
because the file the method is located (bus.util.ts) will cause a
circular references problem when referenced in other classes that are
one of the imports in the bus.util.ts. To prevent any further circular
references issues with this method that's used across many files,
I have moved the main function to util.ts and have the existing method
under bus.util.ts call it.

* Use connString as key to the FabricConnectionState store
* Move connection state store update into broker-connector
* Bump up version to 1.1.0
* Autodetect connString in presence of only one broker
* Update stomp client to usee FabricConnectionStore map