Make getDefaultConnectionString() behavior more predictable
The goal of getDefaultConnectionString() is to look for a connection
string corresponding to the broker target as identified by the params
to `fabric.connect()` from the fabric's internal cache. Previously it
was looking for the available connection string from `connectedMap`
which was updated only after an actual handshake happened. This led to
the behavior of the method unstable because the method grew its
dependency on an asynchronous operation that is session connection.
This MR changes the type of the map it uses to determine the default
connection string. By looking up a connection string in the
`connectionStoreMap`, which gets updated as soon as `fabric.connect()`
is fired, any underlying methods such as `whenConnectionStateChanges()`
can be safely called no matter whether the connection has actually
established.

Signed-off-by: Josh Kim <kjosh@vmware.com>