whenConnectionStateChanges fires properly post fabric connection
This MR fixes an issue with whenConnectionStateChanges() method to
trigger and execute its callback function even after fabric has been
connected beforehand.

Up until now, whenConnectionStateChanges() only fired when a fabric
connection was established after the method had been called.
This was leading to a sub optimal experience with the use of this
method where you needed to be aware of the order of fabric being
connected and this method being invoked. For example, in an Angular app
where it connects to the fabric during the APP_INITIALIZER stage,
navigating from a route to another route that is supposed to invoke the
whenConnectionStateChanges() method would not actually fire and leave
the component hanging indefinitely.

Signed-off-by: Josh Kim <kjosh@vmware.com>