Fix UMD module and update README
UMD module was not usable because of rxjs/operators being not able
to be reached inside the Transport UMD module. This PR builds the
UMD module together with rxjs/operators to fix the issue, at the
cost of a slight increase of bundle size which now sits at 100KB.
Also README got updated with accurate getting started information for
three different ways to import and use the library.

Signed-off-by: Josh Kim <kjosh@vmware.com>