fix: UMD included in bundle after treeshaking
Up until 1.3.3 when importing Transport as an ES6 module the UMD module
file would still be included as part of the application bundle as can
be seen by the breakdown from webpack-bundle-analyzer. This PR fixes it
by removing the `main` entry from the package.json which will result in
about 100kB of reduced bundle size.

Signed-off-by: Josh Kim <jsk9260@gmail.com>